requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
pandas==2.3.3
openai==2.9.0
langchain==1.1.3
langchain-openai==1.1.1
pinecone==8.0.0
rapidfuzz==3.14.3
streamlit==1.52.1
python-dotenv==1.2.1
pydantic==2.12.5
httpx==0.28.1
requests==2.32.5

